refactor: 💡 change type, mechanism id data type

+++ yogidice/src/main/java/com/specialization/yogidice/dto/response/category/MechanismGroupResponse.java
@@ -18,7 +18,7 @@ public class MechanismGroupResponse extends BaseResponseBody {
 
     private String gameTitle;
 
-    private Byte mechanismId;
+    private Long mechanismId;
 
     private String mechanismName;
